refactor: move generic toast formatters to platform/frontend_toasts
- Extract gamepad_connected/disconnected/init and cartridge_load toast
  message functions from src/nes/frontend_toasts.rs into a new
  src/platform/frontend_toasts.rs module (system-agnostic)
- Keep NES-specific emulator_timing and hardware_mode toast functions
  in src/nes/frontend_toasts.rs
- Update imports: event_loop.rs, main.rs use platform::frontend_toasts;
  wasm.rs splits imports between platform (generic) and nes (NES-specific)
- Move corresponding tests alongside the functions
