+++ src/platform/mod.rs
@@ -4,6 +4,7 @@ pub mod config;
 pub mod crc32;
 pub mod debugging;
 pub mod emulator;
+pub mod frontend_toasts;
 pub mod shaders;
 
 #[cfg(feature = "native")]
